+++ packages/web3-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @reef-knot/web3-react
 
+## 1.0.5
+
+### Patch Changes
+
+- Replace Tally legacy connector with Taho wallet adapter
+
 ## 1.0.4
 
 ### Patch Changes